move the date detect function into util and switch to isSameDay func
